Master

+++ helpers/push_helper.py
@@ -5,6 +5,7 @@ from collections import defaultdict
 from google.appengine.ext import db
 from google.appengine.ext import ndb
 from google.appengine.api import users
+from google.appengine.api import memcache
 
 from consts.client_type import ClientType
 from consts.notification_type import NotificationType
@@ -44,11 +45,15 @@ class PushHelper(object):
         workaround for this bug: https://code.google.com/p/googleappengine/issues/detail?id=8848
         solution from: http://stackoverflow.com/questions/816372/how-can-i-determine-a-user-id-based-on-an-email-address-in-app-engine
         """
-        u = users.User(user_email)
-        key = MobileUser(user=u).put()
-        obj = key.get()
-        user_id = obj.user.user_id()
-        key.delete()
+        cache_key = 'PushHelper.user_email_to_id():{}'.format(user_email)
+        user_id = memcache.get(cache_key)
+        if user_id is None:
+            u = users.User(user_email)
+            key = MobileUser(user=u).put()
+            obj = key.get()
+            user_id = obj.user.user_id()
+            key.delete()
+            memcache.set(cache_key, user_id, 60*10)
 
         if Account.get_by_id(user_id) is None:
             # Create an account for this user
@@ -95,17 +100,6 @@ class PushHelper(object):
         return output
 
     @classmethod
-    def get_users_subscribed_to_event(cls, event, notification):
-        keys = []
-        keys.append(event.key_name)
-        keys.append("{}*".format(event.year))
-        users = Subscription.query(Subscription.model_key.IN(keys), Subscription.notification_types == notification).fetch()
-        output = []
-        for user in users:
-            output.append(user.user_id)
-        return output
-
-    @classmethod
     def get_users_subscribed_for_alliances(cls, event, notification):
         keys = []
         for team in event.alliance_teams:
@@ -119,23 +113,20 @@ class PushHelper(object):
 
     @classmethod
     def get_client_ids_for_users(cls, user_list, os_types=None):
+        if not user_list:
+            return defaultdict(list)
+
         if os_types is None:
             os_types = ClientType.names.keys()
-        output = defaultdict(list)
         clients = MobileClient.query(MobileClient.user_id.IN(user_list), MobileClient.client_type.IN(os_types), MobileClient.verified == True).fetch()
+        return cls.get_client_ids_for_clients(clients)
+
+    @classmethod
+    def get_client_ids_for_clients(cls, clients):
+        output = defaultdict(list)
         for client in clients:
             if client.client_type == ClientType.WEBHOOK:
                 output[client.client_type].append((client.messaging_id, client.secret))
             else:
                 output[client.client_type].append(client.messaging_id)
-        return output
-
-    @classmethod
-    def get_all_mobile_clients(cls, client_types=[]):
-        output = []
-        if client_types == []:
-            return output
-        clients = MobileClient.query(MobileClient.client_type.IN(client_types))
-        for user in clients:
-            output.append(user.user_id)
         return output
